+++ lib/types/SmartsheetClient.ts
@@ -1,3 +1,4 @@
+import type { ContactsApi } from '../contacts/types';
 import type { EventsApi } from '../events/types';
 import type { SearchApi } from '../search/types';
 import type { SharingApi } from '../sharing';
@@ -5,7 +6,7 @@ import type { SightsApi } from '../sights/types';
 
 export interface SmartsheetClient {
   constants: any;
-  contacts: any;
+  contacts: ContactsApi;
   events: EventsApi;
   favorites: any;
   folders: any;